fix(api): move error healthcheck on integration test with stub on checkstatuses
